tests: api: use simpler static casts

+++ src/tests/test-api.cc
@@ -95,7 +95,7 @@ TEST_F(ApiTest, Api_TailleRegion)
 
 TEST_F(ApiTest, Api_PositionRegion)
 {
-    EXPECT_EQ(static_cast<unsigned int>(0),
+    EXPECT_EQ(0u,
               players[0].api->positions_region({1, 1}, players[0].id).size());
 
     gs_->set_next_sample({MERCURE, MERCURE});
@@ -114,7 +114,7 @@ TEST_F(ApiTest, Api_PositionRegion)
                     pos.end());
         EXPECT_FALSE(find(pos.begin(), pos.end(), position({2, 2})) !=
                      pos.end());
-        EXPECT_EQ(static_cast<unsigned int>(1), pos.size());
+        EXPECT_EQ(1u, pos.size());
     }
 
     {
@@ -131,17 +131,16 @@ TEST_F(ApiTest, Api_PositionRegion)
                     pos.end());
         EXPECT_TRUE(find(pos.begin(), pos.end(), position({4, 2})) !=
                     pos.end());
-        EXPECT_EQ(static_cast<unsigned int>(5), pos.size());
+        EXPECT_EQ(5u, pos.size());
     }
 }
 
 TEST_F(ApiTest, Api_PlacementsPossibleEchantillon)
 {
-    EXPECT_EQ(
-        static_cast<unsigned int>(2 * 2 * 5 * 6),
-        players[0]
-            .api->placements_possible_echantillon({FER, PLOMB}, players[0].id)
-            .size());
+    EXPECT_EQ(2 * 2 * 5 * 6u, players[0]
+                                  .api->placements_possible_echantillon(
+                                            {FER, PLOMB}, players[0].id)
+                                  .size());
 
     gs_->set_next_sample({MERCURE, MERCURE});
     gs_->reset_turn_state();
@@ -153,34 +152,28 @@ TEST_F(ApiTest, Api_PlacementsPossibleEchantillon)
     gs_->reset_turn_state();
     gs_->place_sample({3, 2}, {4, 2}, players[0].id);
 
-    EXPECT_EQ(static_cast<unsigned int>(40),
-              players[0]
-                  .api->placements_possible_echantillon({MERCURE, MERCURE},
-                                                        players[0].id)
-                  .size());
-    EXPECT_EQ(static_cast<unsigned int>(23),
-              players[0]
-                  .api->placements_possible_echantillon({MERCURE, CUIVRE},
-                                                        players[0].id)
-                  .size());
-    EXPECT_EQ(static_cast<unsigned int>(23),
-              players[0]
-                  .api->placements_possible_echantillon({CUIVRE, MERCURE},
-                                                        players[0].id)
-                  .size());
-    EXPECT_EQ(static_cast<unsigned int>(4),
-              players[0]
-                  .api->placements_possible_echantillon({CUIVRE, CUIVRE},
-                                                        players[0].id)
-                  .size());
-    EXPECT_EQ(
-        static_cast<unsigned int>(0),
-        players[0]
-            .api->placements_possible_echantillon({FER, PLOMB}, players[0].id)
-            .size());
-    EXPECT_EQ(
-        static_cast<unsigned int>(2),
-        players[0]
-            .api->placements_possible_echantillon({FER, CUIVRE}, players[0].id)
-            .size());
+    EXPECT_EQ(40u, players[0]
+                       .api->placements_possible_echantillon({MERCURE, MERCURE},
+                                                             players[0].id)
+                       .size());
+    EXPECT_EQ(23u, players[0]
+                       .api->placements_possible_echantillon({MERCURE, CUIVRE},
+                                                             players[0].id)
+                       .size());
+    EXPECT_EQ(23u, players[0]
+                       .api->placements_possible_echantillon({CUIVRE, MERCURE},
+                                                             players[0].id)
+                       .size());
+    EXPECT_EQ(4u, players[0]
+                      .api->placements_possible_echantillon({CUIVRE, CUIVRE},
+                                                            players[0].id)
+                      .size());
+    EXPECT_EQ(0u, players[0]
+                      .api->placements_possible_echantillon({FER, PLOMB},
+                                                            players[0].id)
+                      .size());
+    EXPECT_EQ(2u, players[0]
+                      .api->placements_possible_echantillon({FER, CUIVRE},
+                                                            players[0].id)
+                      .size());
 }
